+++ src/renderer_d3d12.cpp
@@ -3950,7 +3950,7 @@ namespace bgfx { namespace d3d12
 		uint8_t numStreams = 0;
 		_outNumVertices = _draw.m_numVertices;
 
-		if (UINT8_MAX != _draw.m_streamMask)
+		if (UINT32_MAX != _draw.m_streamMask)
 		{
 			for (uint32_t idx = 0, streamMask = _draw.m_streamMask
 				; 0 != streamMask
@@ -6254,7 +6254,7 @@ namespace bgfx { namespace d3d12
 					const VertexDecl* decls[BGFX_CONFIG_MAX_VERTEX_STREAMS];
 
 					uint8_t numStreams = 0;
-					if (UINT8_MAX != draw.m_streamMask)
+					if (UINT32_MAX != draw.m_streamMask)
 					{
 						for (uint32_t idx = 0, streamMask = draw.m_streamMask
 							; 0 != streamMask
